+++ CHANGELOG.md
@@ -73,6 +73,7 @@ we hit release version 1.0.0.
   be done for assigning matrix elements (it fills with 0's).
 
 ### Removed
+- `cell` argument for `Geometry.translate/move` (it never worked)
 - removed `Selector` and `TimeSelector`, they were never used internally
 
 ### Changed